schema stuff

+++ ToyDb/Schema.cs
@@ -6,7 +6,7 @@ public class Schema(string name)
     public List<Field> Fields { get; set; }
 
     // todo: should not need length except for strings 
-    public Schema AddField(string name, SchemaFieldType type, int length)
+    public Schema AddField(string name, SchemaFieldType type, byte length)
     {
         if (length == 0) throw new ArgumentNullException(nameof(length));
         
@@ -15,12 +15,12 @@ public class Schema(string name)
     }
 }
 
-public class Field(string name, SchemaFieldType type, int offset, int length)
+public class Field(string name, SchemaFieldType type, int offset, byte length)
 {
     public string Name { get; set; } = name;
     public SchemaFieldType Type { get; set; } = type;
     public int Offset { get; set; } = offset;
-    public int Length { get; set; } = length;
+    public byte Length { get; set; } = length;
 }
 
 public enum SchemaFieldType
